Refactor modificando tipo de dato cc de Integer por Long

+++ src/main/java/com/example/Banco/Dto/UsuarioDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class UsuarioDTO {
     private String nombre;
     private String apellido;
-    private Integer cedula;
+    private Long cedula;
     private String correo;
     private String password;
 }
